Converted over to using MySQL instead of SQL Server Updated various third party lib versions Updated favicon Switched from DapperExtensions to much simpler SimpleCRUD and embedded a version which I've tweaked due to a change in the returned type of LAST_INSERT_ID() which breaks both Dapper and SimpleCRUD. Given SimpleCRUD is a much simpler file it was easier to patch up than DapperExtensions, which I'm not using the majority of anyway. Also changed SimpleCRUD to use MySQL syntax instead of that for SQL Server. Updated Author page to internalise some code for a deleted user.

+++ Source/Models/DebateLinkVote.cs
@@ -6,7 +6,7 @@ namespace RationalVote.Models
     public partial class DebateLinkVote
     {
         public long Id { get; set; }
-        public short Vote { get; set; }
+        public byte Vote { get; set; }
     
         public virtual DebateLink Link { get; set; }
         public virtual User Owner { get; set; }
